requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.47.0
blessed==1.21.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
bypy==1.8.9
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
cfgv==3.4.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
colour-demosaicing==0.2.6
colour-science==0.4.6
contourpy==1.3.2
cupy @ file:///home/conda/feedstock_root/build_artifacts/cupy-split_1752439477717/work
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
distlib==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
facexlib==0.3.0
fastrlock @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastrlock_1734873426/work
filelock @ file:///croot/filelock_1744281381737/work
filterpy==1.4.5
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gpustat==1.1.1
grpcio==1.74.0
hf-xet==1.1.7
huggingface-hub==0.34.4
icecream==2.1.5
identify==2.6.13
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
iniconfig==2.1.0
Jinja2 @ file:///croot/jinja2_1741710844255/work
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.7.3
lpips==0.1.4
Markdown==3.8.2
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.5
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.6.4
multiprocess==0.70.16
networkx @ file:///croot/networkx_1737039604450/work
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-clip==1.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow @ file:///croot/pillow_1752524526621/work
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2025.7.33
requests @ file:///croot/requests_1750426329888/work
requests-toolbelt==1.0.0
ruff==0.12.8
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sympy @ file:///croot/sympy_1738108488918/work
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.15.2
tomli==2.2.1
torch==2.8.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.37.2
triton==3.4.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
virtualenv==20.33.1
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
